backend/requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Web Scraping & HTTP Requests
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0

# HTML/XML Parsing
html5lib>=1.1
pyquery>=2.0.0

# Date/Time Handling
python-dateutil>=2.8.0

# Text Processing & NLP
nltk>=3.8
spacy>=3.7.0
textblob>=0.17.0

# AI/ML Libraries
openai>=1.0.0
anthropic>=0.7.0
transformers>=4.35.0
torch>=2.0.0

# Data Validation & Models
pydantic>=2.4.0
marshmallow>=3.20.0

# Database & Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.5.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0
configparser>=5.3.0

# Scheduling & Task Management
schedule>=1.2.0
celery>=5.3.0
redis>=5.0.0

# Email & Notifications
secure-smtplib
email-validator>=2.1.0

# Logging & Monitoring
loguru>=0.7.0

# File Handling
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# HTTP Server (if needed for API)
fastapi>=0.104.0
uvicorn>=0.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
requests-mock>=1.11.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0

# Utilities
tqdm>=4.66.0
click>=8.1.0
pathlib2>=2.3.0
